Move `channelScalePropertyIncompatability` to `src/scale` as it can be reused in other method such as CQL
