+++ src/compile/scale/range.ts
@@ -3,12 +3,10 @@ import * as log from '../../log';
 import {COLUMN, ROW, X, Y, SHAPE, SIZE, COLOR, OPACITY, Channel} from '../../channel';
 import {Config} from '../../config';
 import {Mark} from '../../mark';
-import {Scale, ScaleConfig, ScaleType, scaleTypeSupportProperty, isExtendedScheme} from '../../scale';
+import {Scale, ScaleConfig, ScaleType, channelScalePropertyIncompatability, scaleTypeSupportProperty, isExtendedScheme} from '../../scale';
 import {Type} from '../../type';
 import {VgRange, VgRangeScheme} from '../../vega.schema';
 import * as util from '../../util';
-
-import {channelScalePropertyIncompatability} from './scale';
 
 export type RangeMixins = {range: VgRange} | {rangeStep: number};
