+++ src/compile/scale/scale.ts
@@ -1,10 +1,5 @@
 import {VgScale} from '../../vega.schema';
 
-import {Channel} from '../../channel';
-import * as log from '../../log';
-
-
-
 /** Scale suffix for scale used to get drive binned legends. */
 export const BIN_LEGEND_SUFFIX = '_bin_legend';
 /** Scale suffix for scale for binned field's legend labels, which maps a binned field's quantitative values to range strings. */
@@ -22,51 +17,3 @@ export type ScaleComponents = {
   binLegendLabel?: ScaleComponent;
 };
 
-export function channelScalePropertyIncompatability(channel: Channel, propName: string): string {
-  switch (propName) {
-    case 'range':
-      // User should not customize range for position and facet channel directly.
-      if (channel === 'x' || channel === 'y') {
-        return log.message.CANNOT_USE_RANGE_WITH_POSITION;
-      }
-      if (channel === 'row' || channel === 'column') {
-        return log.message.cannotUseRangePropertyWithFacet('range');
-      }
-      return undefined; // GOOD!
-    // band / point
-    case 'rangeStep':
-      if (channel === 'row' || channel === 'column') {
-        return log.message.cannotUseRangePropertyWithFacet('rangeStep');
-      }
-      return undefined; // GOOD!
-    case 'padding':
-    case 'paddingInner':
-    case 'paddingOuter':
-      if (channel === 'row' || channel === 'column') {
-        /*
-         * We do not use d3 scale's padding for row/column because padding there
-         * is a ratio ([0, 1]) and it causes the padding to be decimals.
-         * Therefore, we manually calculate "spacing" in the layout by ourselves.
-         */
-        return log.message.CANNOT_USE_PADDING_WITH_FACET;
-      }
-      return undefined; // GOOD!
-    case 'scheme':
-      if (channel !== 'color') {
-        return log.message.CANNOT_USE_SCHEME_WITH_NON_COLOR;
-      }
-      return undefined;
-    case 'type':
-    case 'domain':
-    case 'round':
-    case 'clamp':
-    case 'exponent':
-    case 'nice':
-    case 'zero':
-    case 'useRawDomain':
-      // These channel do not have strict requirement
-      return undefined; // GOOD!
-  }
-  /* istanbul ignore next: it should never reach here */
-  throw new Error('Invalid scale property "${propName}".');
-}
