+++ src/scale.ts
@@ -1,3 +1,5 @@
+import * as log from './log';
+import {Channel} from './channel';
 import {DateTime} from './datetime';
 import {contains, toSet} from './util';
 
@@ -307,4 +309,56 @@ export function scaleTypeSupportProperty(scaleType: ScaleType, propName: string)
   }
   /* istanbul ignore next: should never reach here*/
   throw new Error(`Invalid scale property ${propName}.`);
+}
+
+/**
+ * Returns undefined if the input channel supports the input scale property name
+ */
+export function channelScalePropertyIncompatability(channel: Channel, propName: string): string {
+  switch (propName) {
+    case 'range':
+      // User should not customize range for position and facet channel directly.
+      if (channel === 'x' || channel === 'y') {
+        return log.message.CANNOT_USE_RANGE_WITH_POSITION;
+      }
+      if (channel === 'row' || channel === 'column') {
+        return log.message.cannotUseRangePropertyWithFacet('range');
+      }
+      return undefined; // GOOD!
+    // band / point
+    case 'rangeStep':
+      if (channel === 'row' || channel === 'column') {
+        return log.message.cannotUseRangePropertyWithFacet('rangeStep');
+      }
+      return undefined; // GOOD!
+    case 'padding':
+    case 'paddingInner':
+    case 'paddingOuter':
+      if (channel === 'row' || channel === 'column') {
+        /*
+         * We do not use d3 scale's padding for row/column because padding there
+         * is a ratio ([0, 1]) and it causes the padding to be decimals.
+         * Therefore, we manually calculate "spacing" in the layout by ourselves.
+         */
+        return log.message.CANNOT_USE_PADDING_WITH_FACET;
+      }
+      return undefined; // GOOD!
+    case 'scheme':
+      if (channel !== 'color') {
+        return log.message.CANNOT_USE_SCHEME_WITH_NON_COLOR;
+      }
+      return undefined;
+    case 'type':
+    case 'domain':
+    case 'round':
+    case 'clamp':
+    case 'exponent':
+    case 'nice':
+    case 'zero':
+    case 'useRawDomain':
+      // These channel do not have strict requirement
+      return undefined; // GOOD!
+  }
+  /* istanbul ignore next: it should never reach here */
+  throw new Error('Invalid scale property "${propName}".');
 }
